# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Darbby Platform API — Backend for customer and merchant apps
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trips:
    get:
      operationId: listMyTrips
      tags: [trips]
      summary: List my trips
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [ACTIVE, COMPLETED, CANCELLED]
      responses:
        "200":
          description: List of trips
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trip"
    post:
      operationId: createTrip
      tags: [trips]
      summary: Create a new trip
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTripRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
  /trips/{id}:
    get:
      operationId: getTripById
      tags: [trips]
      summary: Get trip details with matched offers
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UuidId"
      responses:
        "200":
          description: Trip with offers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripDetail"
    patch:
      operationId: updateTripStatus
      tags: [trips]
      summary: Update trip status
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UuidId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [COMPLETED, CANCELLED]
              required: [status]
      responses:
        "200":
          description: Updated trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
  /trips/{id}/nearby-merchants:
    get:
      operationId: getNearbyMerchants
      tags: [trips]
      summary: Get merchants whose branches intersect with the trip route
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UuidId"
      responses:
        "200":
          description: Nearby merchant branches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NearbyBranch"

  # ─── OFFERS ───────────────────────────────────────────────────────────────
  /offers/{id}:
    get:
      operationId: getOffer
      tags: [offers]
      summary: Get offer details
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UuidId"
      responses:
        "200":
          description: Offer with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OfferDetail"
  /offers/{id}/accept:
    post:
      operationId: acceptOffer
      tags: [offers]
      summary: Accept an offer (customer)
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UuidId"
      responses:
        "200":
          description: Offer accepted and commission ledger entry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OfferDetail"
  /offers/{id}/reject:
    post:
      operationId: rejectOffer
      tags: [offers]
      summary: Reject an offer (customer)
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UuidId"
      responses:
        "200":
          description: Offer rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OfferDetail"
  /offers/{id}/counter:
    post:
      operationId: counterOffer
      tags: [offers]
      summary: Submit a counter-offer (negotiation round)
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UuidId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CounterOfferRequest"
      responses:
        "200":
          description: Counter-offer submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Negotiation"

  # ─── MERCHANT OFFERS ──────────────────────────────────────────────────────
  /merchant/trips:
    get:
      operationId: listActiveTripsMerchant
      tags: [merchants]
      summary: List active trips near merchant branches
      security:
        - bearerAuth: []
      parameters:
        - name: branchId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Matching trips
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trip"
...
  /merchant/offers/{id}/counter:
    post:
      operationId: merchantCounter
      tags: [merchants]
      summary: Merchant sends a counter-offer
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UuidId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CounterOfferRequest"
      responses:
        "200":
          description: Counter submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Negotiation"

  # ─── MERCHANT PROFILE ─────────────────────────────────────────────────────
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List my notifications
      security:
        - bearerAuth: []
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UuidId"
      responses:
        "204":
          description: Marked as read

  # ─── GEO / MAPS ──────────────────────────────────────────────────────────
  /geo/route:
    post:
      operationId: getRoute
      tags: [geo]
      summary: Get route polyline and distance from Google Maps
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteRequest"
      responses:
        "200":
          description: Route information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteResponse"

  # ─── MERCHANT STATS ───────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UuidId:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    CreateTripRequest:
      type: object
      properties:
        title:
          type: string
        tripPurpose:
          type: string
          enum: [WORK, TOURISM, UMRAH, FAMILY_VISIT, MEDICAL, EDUCATION, OTHER]
        originName:
          type: string
        originLat:
          type: number
        originLng:
          type: number
        destinationName:
          type: string
        destLat:
          type: number
        destLng:
          type: number
        routePolyline:
          type: string
        departureTime:
          type: string
          format: date-time
        arrivalTime:
          type: string
          format: date-time
        vehicleProfileId:
          type: string
        isPublic:
          type: boolean
      required: [originName, originLat, originLng, destinationName, destLat, destLng, departureTime, tripPurpose]
    Trip:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        vehicleProfileId:
          type: string
        title:
          type: string
        tripPurpose:
          type: string
        originName:
          type: string
        destinationName:
          type: string
        routePolyline:
          type: string
        departureTime:
          type: string
          format: date-time
        arrivalTime:
          type: string
          format: date-time
        status:
          type: string
        isPublic:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, userId, tripPurpose, originName, destinationName, departureTime, status]
    TripDetail:
      allOf:
        - $ref: "#/components/schemas/Trip"
        - type: object
          properties:
            offers:
              type: array
              items:
                $ref: "#/components/schemas/OfferDetail"
    NearbyBranch:
      type: object
      properties:
        branchId:
          type: string
        merchantId:
          type: string
        businessName:
          type: string
        branchName:
          type: string
        distanceMeters:
          type: number
        addressText:
          type: string
        serviceRadiusKm:
          type: number
      required: [branchId, merchantId, businessName, branchName]

    # ─── Offer ───────────────────────────────────────────────────────────────
...
    CounterOfferRequest:
      type: object
      properties:
        proposedPrice:
          type: number
        message:
          type: string
      required: [proposedPrice]
    OfferDetail:
      type: object
      properties:
        id:
          type: string
        tripId:
          type: string
        merchantId:
          type: string
        branchId:
          type: string
        message:
          type: string
        totalPrice:
          type: number
        finalPrice:
          type: number
        status:
          type: string
        isAutoOffer:
          type: boolean
        expiresAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/OfferItemDetail"
        negotiations:
          type: array
          items:
            $ref: "#/components/schemas/Negotiation"
        createdAt:
          type: string
          format: date-time
      required: [id, tripId, merchantId, totalPrice, status]
    OfferItemDetail:
      type: object
      properties:
        id:
          type: string
        productId:
          type: string
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        negotiatedPrice:
          type: number
      required: [id, productId, quantity, unitPrice]
    Negotiation:
      type: object
      properties:
        id:
          type: string
        offerId:
          type: string
        senderType:
          type: string
          enum: [USER, MERCHANT, SYSTEM]
        proposedPrice:
          type: number
        message:
          type: string
        isAuto:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, offerId, senderType, proposedPrice]
...
    Notification:
      type: object
      properties:
        id:
          type: string
        recipientType:
          type: string
        recipientId:
          type: string
        type:
          type: string
        title:
          type: string
        body:
          type: string
        data:
          type: object
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, recipientType, recipientId, type, title, body, isRead]

    # ─── Geo / Maps ───────────────────────────────────────────────────────────
    RouteRequest:
      type: object
      properties:
        originLat:
          type: number
        originLng:
          type: number
        destLat:
          type: number
        destLng:
          type: number
      required: [originLat, originLng, destLat, destLng]
    RouteResponse:
      type: object
      properties:
        polyline:
          type: string
        distanceMeters:
          type: integer
        durationSeconds:
          type: integer
        distanceText:
          type: string
        durationText:
          type: string
      required: [polyline, distanceMeters, durationSeconds]

    # ─── Stats ────────────────────────────────────────────────────────────────
...