# lib/api-zod/src/generated/types/routeRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Darbby Platform API — Backend for customer and merchant apps
 * OpenAPI spec version: 1.0.0
 */

export interface RouteRequest {
  originLat: number;
  originLng: number;
  destLat: number;
  destLng: number;
}